Rename tc-text to chat-message-text
