+++ assets/javascripts/discourse/templates/components/chat-message-text.hbs
