+++ assets/stylesheets/common/common.scss
@@ -1342,7 +1342,7 @@ body.composer-open .topic-chat-float-container {
 }
 
 .tc-message-collapsible,
-.tc-text {
+.chat-message-text {
   > p {
     margin: 0.5em 0 0.5em;
   }
