+++ assets/stylesheets/desktop/desktop.scss
@@ -61,7 +61,7 @@
   padding: 0.65em 1em 0.15em;
 }
 
-.tc-text {
+.chat-message-text {
   img:not(.emoji):not(.avatar) {
     -webkit-transition: all 0.6s cubic-bezier(0.165, 0.84, 0.44, 1);
     transition: all 0.6s cubic-bezier(0.165, 0.84, 0.44, 1);
